Revert "Unprompt channels during project load"
This reverts commit b1dcf031a585d8f6ba8cc86f3e215c9673f9bd18.

I found that b1dcf031a585d8f6ba8cc86f3e215c9673f9bd18 had some
unintended consequences that seemed to mess up the prompt state. The
real problem that it was trying to address was that the prompt was being
interleaved with log messages in some scenarios. There was a different
way to fix that in ProgressState that was both simpler and more
reliable.